chore: improve project tooling (#513)
## Tasks

This pull request makes some changes to improve precise details on our
tooling

### `.lintstagedrc`
- [x] Change to `prettier --write` and `eslint --fix`. So, instead of
just checking on husky, it will format and fix the files before
committing.
- [x] Add other formats on `lintstaged` to run `prettier`

### `.prettierrc`
- [x] Setup prettier to format all types of files, since we have a bunch
new of them and this configuration came from `fuels-ts` that's a totally
different type of project

### `.vscode` folder
- [x] Add a recommendation of a new extension that is a [Jest
runner](https://marketplace.visualstudio.com/items?itemName=firsttris.vscode-jest-runner)
inside VSCode. I used this extension on `fuel-ui` and it's capable to
debug using VSCode and Jest, it's very useful for debugging without
console.log on tests, and this extension is way better than the with
most downloads on VSCode extensions store.