fix: refactor on store to fix bugs related to it + improvements (#544)
*tdlr;*

⭐ Fix #376 #531 

# Introduction
This pull request aims to improve the current global xstate machines
system (aka xstore) in our application. The current system is facing
several problems such as cycle dependencies, unexpected bugs, and
difficulty in tracking and debugging issues. To overcome these
challenges, a new approach has been proposed using Jotai as a global
store to manage the machines outside the React scope. The new approach
not only brings performance improvements but also reduces the chances of
encountering unexpected bugs.

## Main problems
- Cycle dependencies are causing difficulties in tracking issues in the
application.
- Unexpected bugs related to runtime variables are arising due to the
recreation of classes on live-reloading.
- Services are being started multiple times leading to performance
issues.
- The current non-reactive implementation is making it hard to catch
bugs.

## Solution
- To solve these problems, a new `StoreProvider` has been introduced in
the application using [Jotai](https://jotai.org/) behind the scenes as a
global store to provide a safer and more robust solution, giving more
control and predictability. Additionally, the new approach has
introduced several APIs to the store and improved the overall
performance of the application and now localStorage is used for cache
state with an option to blacklist a state.

## New APIs
##### `store.addMachine(key, () => machine)`
This API allows adding a new state machine to the store.

##### `store.addHandlers((store: Store) => Handlers)`
This API allows adding handlers for state management.

##### `store.onStateChange(key | Service, (state: State) => void)`
This API allows listening to changes in state.

##### `store.onStartStore(() => void)`
This API allows executing logic when the store starts.

##### `store.getStateFrom(key)`
This API allows get some state from a specific service

##### `store.waitFor(key, string | (state) => boolean)`
This API allows waiting for some service state using the store

##### `const [state, send] = store.useState('accounts')`
This API allows using the state of the store in components.

## LocalStorage cache

As mentioned now it's possible to have a cache in some states using
LocalStorage to get the initial state when service starts. This also
possible to control which states you want to be saved or not by using
`storageBlacklist` prop:

### Usage

```tsx
import { createStore } from '@fuel-wallet/xstore';

export enum Services {
  accounts = 'accounts',
  networks = 'networks',
}

export type StoreMachines = {
  accounts: AccountMachine;
  networks: NetworksMachine;
};

export const store$ = createStore<StoreMachines>({
  id: 'fuelStore',
  persistedStates: ['networks'],
});

export const store = store$
  .addMachine(Services.accounts, () => accountMachine)
  .addMachine(Services.networks, () => networksMachine)
  .addHandlers(accountEvents)
  .addHandlers(networkEvents)
  .setup();

export const { StoreProvider } = store;
```

## Pros/Cons to have a global store

### Pros
- The integration of these APIs enhances the application's performance
and simplifies state management, reducing the risk of errors. This leads
to increased predictability and fewer unexpected bugs.
- Be able to send a dispatch from the store anywhere and outside React
context
- Be able to make a broadcast and send an event for all machines
- Simple API to handle things
- Be able to `useSetConfigMachine`
- Don't need to create a dependency between machines (decoupling)

### Cons
- We have a trade-off right now that's a new library in the lower layer
of the application. In fact, we could choose to try to implement all in
native Javascript *(I tried hard, believe me)* but after days of working
on this, was hard to find a better/safer solution that covered all
things we need.
- Abstractions that can lead to bugs sometimes

Since [Jotai has just
2.3kb](https://bundlephobia.com/package/jotai@2.0.0) gzip, I don't see
this as a big problem for us. I think is worth the trade-off.

> 📝 before someone asks, we have technical requirements that prevent us
from using React.Context. Instead, we are able to dispatch actions from
the store anywhere in the application by simply importing it and calling
`store.send`. This approach would not be possible with React Context, as
we would need to constantly inject actions into the state machines.