fix: react hooks observers (#1127)
This is exactly the same implementation of
https://github.com/FuelLabs/fuels-npm-packs/pull/55
Applying here just for compatibility purposes (temporarily since this
will be removed soon).

-----

This PR fixes and reduces a lot of re-renders across our query
observers.

This issue was caused due to the way which `@tanstack/react-query`
handles the tracked properties to re-renders.
Internally, they implemented a `Proxy` to intercept and track which
properties a component is listening from a `useQuery` hook.

It means that when we're calling a `useQuery` with spread operators, it
is forwarding to that `Proxy` that we want to listen to EVERY existent
property.

 ```tsx
const query = useQuery();

return { custom: query.data, ...query }; // <- "...query" is the issue
here
```

So, what I did? I created the `useNamedQuery` that is also a `Proxy` and it'll also rename the data property to anything you want to name, and keeping the performance everywhere.

This ensures we're going to re-render **ONLY** when `tracked` properties get updates. 


| ⚠️  Before |  ✅ Now |
| ------ | ------ |
| <img width="288" alt="Screenshot 2024-03-08 at 11 13 38" src="https://github.com/FuelLabs/fuels-npm-packs/assets/7074983/f5aabad7-60ec-4674-bd8a-8b6280c375ef"> | Only 1 property will trigger re-renders <img width="264" alt="Screenshot 2024-03-08 at 11 12 59" src="https://github.com/FuelLabs/fuels-npm-packs/assets/7074983/30ad77ee-399f-44cf-a99f-6af36ae6a0c0"> |