ci(changesets): versioning packages (#1259)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to master, this PR
will be updated.


# Releases
## fuels-wallet@0.16.5

### Patch Changes

- [#1255](https://github.com/FuelLabs/fuels-wallet/pull/1255)
[`4b7d4bd3`](https://github.com/FuelLabs/fuels-wallet/commit/4b7d4bd398448259a8eeb21bf7268df92891ad39)
Thanks [@luizstacio](https://github.com/luizstacio)! - fix: auto-update
in background to wait for update to be downloaded

-   Updated dependencies \[]:
    -   @fuel-wallet/connections@0.16.5

## @fuel-wallet/connections@0.16.5



## @fuel-wallet/types@0.16.5

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>