chore(e2e): Added support for the 'chrome-dev' channel in Playwright FE-803 (#1677)
~~Add~~ Replace `chromium` with `chrome-beta` tests as a way to catch
breaking changes early on.

I would have chosen the `dev` channel as it gives us 9 to 12 weeks to
catch those changes, as the `beta` channel would only give 4 to 6 weeks
and the `Canary` version is too unstable, but `dev` is not available on
playwright at the moment.

Decided to replace to the beta version instead of running on both as to
not increase the test time. If it runs on beta we can be safe that it
runs on stable.