chore (#1834)
<!--
List the issues that this PR closes using a bullet list. For example:
- Closes #X
- Closes FE-Z
-->

<!--
# Summary
Please provide a summary of your changes if they are not
self-explanatory from the closed issues. In many cases, this section can
be removed if the issues fully explain the reason for this PR.
-->

# Checklist

- [ ] I've added error handling for all actions/requests, and verified
how this error will show on UI. (or there was no error handling)
- [ ] I've reviewed all the copy changed/added in this PR, using AI if
needed. (or there was no copy changes)
- [ ] I've included the reference to the issues being closed from Github
and/or Linear (or there was no issues)
- [ ] I've changed the Docs to reflect my changes (or it was not needed)
- [ ] I've put docs links where it may be helpful (or it was not needed)
- [ ] I checked the resulting UI both in Light and Dark mode (or no UI
changes were made)
- [ ] I **reviewed** the **entire PR** myself (preferably, on GH UI)